// server/src/types/newsApi/newsItemSummaryResponse.ts
export interface NewsItemSummaryResponse {
	response: {
		status: string;
		userTier: string;
		total: number;
		startIndex: number;
		pageSize: number;
		currentPage: number;
		pages: number;
		orderBy: string;
		results: Result[];
	};
}

interface Result {
	id: string;
	type: string;
	sectionId: string;
	sectionName: string;
	webPublicationDate: string;
	webTitle: string;
	webUrl: string;
	apiUrl: string;
	isHosted: boolean;
	pillarId: string;
	pillarName: string;
	fields: {
		headline: string;
		shortUrl: string;
		thumbnail: string;
	};
}